Plugin Directory: Allow authors to update their plugin submissions.
If the initial version of a plugin does not satisfy review requirements, plugin
authors need a way  to update their plugin for further review.
This now adds subsequent zip files by the same author to existing tickets and
adjusts logic and UI in a few places to handle multiple zip files.

See #1570.


git-svn-id: https://meta.svn.wordpress.org/sites/trunk@3034 74240141-8908-4e6f-9713-ba540dce6ec7